+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2025-08-13
+
+Windows support! Mostly worked on Windows already but this update
+makes timeouts work properly, properly locates the `config.json` file
+and updates the documentation.
+
+
 ## [0.11.0] - 2025-08-11
 
 Added "quiet mode" -q to eliminate noise, for use in testing environments.